perf: reduce read-path memory copies
Every full-data representation on the PLY read path is now released as
soon as the next one is built, and the biggest transient — the USDA
serialization and reparse round-trip — is gone:

- PlyReader decodes properties directly to float (out-of-float-range
  doubles become +/-infinity so finiteness validation still rejects
  them) and frees each tinyPLY native buffer as its column converts.
- GaussianPlyDecoder decodes column-wise, moving each property column
  out of the PlyDocument and freeing it as it enters GaussianCloudData,
  replacing the per-row map-lookup loop.
- GaussianLayerWriter consumes the cloud (WriteToLayer takes it by
  rvalue), releases each source array once copied into its COW VtArray
  attribute, and returns the authored anonymous layer directly.
- GaussianPlyFileFormat::Read transfers the worker-authored layer with
  TransferContent. The worker-thread SdfChangeBlock workaround (design
  policy 7.5) is unchanged.

Measured on local Postshot captures (SH degree 3), peak commit /
Usd.Stage.Open before -> after:
  139k Gaussians: 1.70 GiB / 3.1 s -> 0.18 GiB / 0.4 s
  696k Gaussians: 8.05 GiB / 15.3 s -> 0.40 GiB / 1.9 s
  1.94M Gaussians (after only): 0.92 GiB / 5.2 s

Authored stage content is unchanged except the layer doc no longer
embeds the doubled round-trip provenance line; the L5 golden is
regenerated accordingly. Workspace ost plugin test --up-to 5: 12 pass /
0 fail / 3 skip; CTest 2/2.

+++ plugins/gaussian-ply/src/io/PlyReader.h
@@ -34,12 +34,14 @@ struct PlyHeader {
 
 struct PlyDocument {
     PlyHeader header;
-    std::map<std::string, std::vector<double>> vertexProperties;
+    std::map<std::string, std::vector<float>> vertexProperties;
 };
 
 // Isolates tinyPLY from the semantic decoder. The adapter exposes only scalar
-// vertex properties converted to double so the rest of the importer does not
-// depend on tinyPLY's buffers or type enum.
+// vertex properties converted to float so the rest of the importer does not
+// depend on tinyPLY's buffers or type enum. Values a float cannot represent
+// become +/-infinity so the decoder's finiteness validation still rejects
+// out-of-range source data.
 class PlyReader {
 public:
     bool ReadHeader(
